Add teams dataset pipeline
Used to import teams data from datahub. One thing to note is that
currently datahub endpoint is only ordered by ID (which is a UUID),
so it's possible that pagination will break (by including an item
more than once) if a new team is created during the import. Teams
don't have a creation date column, so there's no other way to order
them at the moment, but the table shouldn't change too often.