Apply autoformatting changes
This is the result of running `make format` for the first time. No
other changes have been made, so the code should stay functionally
the same.

black seems to join a lot of lines in dataset_pipelines, which makes
it slightly harder to to scan through. I'm on the fence about this,
but still think the benefits of standardised auto-formatting outweigh
the individual style points. Happy to drop this for now if people
disagree though.