Fix global variable reference in mark_task_failed
Now that the DAG creation has been moved to a separate module flake8
recognizes that mark_task_failed was using global run_fetch_task_id,
which was defined in the DAG creation loop (so it would always end up
with the ID for the last fetch task created during import).

This fixes the issue by passing run_fetch_task id from the fetch task
and only cleaning up state when it's set. So a failed fetch task will
delete redis cache / variables that it sets, but other tasks following
it would not.