Move airflow PID file to /tmp directory when running inside Docker
Airflow saves the PID file in AIRFLOW_HOME by default, which when
running inside a Docker container locally is mapped to a local
folder, which means it can sometimes remain after container was
shutdown/killed, preventing webserver from starting up on the next
run.

Moving it to /tmp makes sure the PID file is lost with the process
when the container is restarted.