Add new SQLAlchemy-based implementations for DB dataset operators
This reimplements existing dataset operators with a few changes:

Operators are now using SQLAlchemy to perform most DB operations
like creating and dropping tables and inserting data. This requires
the DAG/pipeline to create an SQLAlchemy Table instance that is used
as a base for all operations. As a result, we can avoid most SQL
templates and rely on SQLAlchemy to generate create table and insert
statements as well as convert python data types to SQL where necessary.

Operators are designed to run sequentially within a DAG, not in parallel.
This gets rid of `mark_task_failed` and XCom checks for current fetch
run status since the fetch task should have finished successfully before
the DB operators are started. This means we can't run fetch and import
at the same time (potentially increasing overall DAG run time), but
makes the pipeline and operators much simpler. The import task now runs
only when fetched has successfully completed and can go through the full
list of pages without waiting or dealing with a queue. It fixes some
bugs with tasks not cleaning up correctly or not being able to retry
(like not restoring the page back to the redis queue after failing,
therefore losing a page of results).

The "move data from temporary table" operator is re-implemented using
table renames instead of TRUNCATE and INSERT. This reduces the amount
of time the dataset is unavailable and makes it possible to migrate
dataset schema with each run (since a fresh new table is created), but
currently won't work with any data sets that have a data cut view based
on them since the views will still be tied to the old (now renamed)
table. This can be addressed by e.g. including view generation into the
overall pipeline in the future, so that each time a dataset is
reimported views are redefined before the tables are swapped.