Rewrite SSO login to use requests-oauthlib instead of Flask-OAuthlib
Flask-OAuthlib is deprecated. The repository is archived and it hasn't
been updated in 11 months prior to that. It also has a dependency
conflict with requests-oauthlib, which it requires, making it impossible
to do a strict dependency resolution for data-flow.

This rewrites the SSO login module to use the underlying requests-oauthlib
instead.

This also fixes a bug in `domain_check` where an email address could pass
the check by including`"@<allowed_domain>@..."` in the local part of the
address (which is allowed by the spec).