Pin and upgrade all application requirements
This moves our dependencies list to `requirements.in` and uses
`pip-compile` to generate a full requirements.txt, listing all direct
and third-party requirements.

requirements.in are unpinned since we don't have any known
incompatibilities with the latest package versions, so as a result
all packages have been upgraded to their latest version.