Release a specific commit through all CI stages
Promotion pipeline should consistently move the same commit through
dev, staging and prod, even if there were new merges into master
since the job has initially been started.

The easiest way to set a variable that is available in all stages of
the Jenkins Pipeline and doesn't require an agent to access seems to
be to set a parameter when starting the build. This changes the job
to accept GIT_COMMIT parameter, checkout that specific commit to
populate build name with a PR number and then call ci-pipeline with
it.

In order to set the commit parameter automatically on each master merge
this job will be started by another Jenkins job, which watches the Github
repo.