Split large ONS datasets by indexing over one of the dimensions
This implements the "cursoring" approach proposed by ONS: we split
the dataset into smaller ones by first getting a list of all possible
values for one or more dimensions (e.g. geography or year/month) and
then requesting data for each value (or combination of values) of these
dimensions.

In order to achieve this we add an optional `index_query` to the ONS
pipeline class. This query returns a list of values used to parameterize
the dataset query in order to filter out results. The query should also
return a `label` value, which is used to prefix page response file names
stored in S3.

Index query is currently not paginated. With a raised default page limit
of 10000, a paginated response would require us to make more than 10000
HTTP requests, so we might need a different approach in that case.