Split ONS UK trade data dataset import by the time period
ONS UK traded data contains 15 million records and is too large to
paginate through using LIMIT and OFFSET.

This is using the `index_query` to retrieve a list of time periods
used by the dataset (~250 values, one for each month since 1998)
and then request data for each of those periods separately. With a
page size of 10000 this results in 6 requests per month, for a total
of ~1500 requests, which take around 1.5 hours to import and store
in S3.

Since we're using `str.format` to inject the filter value into the
query, there's a small escape sequence conflict between literal `{`
required by SPARQL `WHERE` clause escaped as `{{` for python string
formatting and Airflow interpreting parameters as Jinja2 templates
by default. To avoid having `{{ ...  }}` interpreted as Jinja2 tag
we wrap part of the query with Jinja `{% raw %}` tag.