Create new base classes for pipeline and CSV DAGs
This merges the changes from different Base...Pipeline classes into
one common base class that allows overriding the fetch operator.

This reduces the duplication of the common DAG structure and makes
sure all datasets are always running the same operators.

This also replaces the `__subclass__` DAG registration loop with a
metaclass that creates a DAG module-level variable on import.

The main benefit of the metaclass is that it doesn't require importing
all modules defining subclasses for the given base class, which makes
it much easier to re-use the same base class across modules.

_PipelineDAG is named to work around airflow's "safe mode" import,
which requires strings "airflow" and "DAG" to be present in the
file contents.