Add a dataflow logger for operator logs
Airflow upgrade changed something in the logging configuration so
our operator log messages were no longer included in the task logs.

This creates and configures a new `dataflow` logger that should be
used by all operators for logs to be stored and shipped to S3 with
the airflow task logs.