Remove HMRC validators to parse data directly from list
HMRC export and import data is provided as |-delimited text files.
The files have a strictly defined format, so instead of trying to
validate them against it we parse the file as defined and rely on
DB column constraints to raise an error if a row happens to be
invalid in the future.

This is the first dataset we fetch that is not in JSON format, which
is what S3Data expects. We could trivially extend it to dump/load
arbitrary files, but this requires more coordination between fetch
and DB insert tasks (like knowing that non-JSON files are line based,
can have headers and footers that might need to be ignored etc.).
It's probably worth doing once we have more examples, but for now
we use fetch to transform text file into a JSON list of lists, remove
header and footer lines and store the result.