Add a transform to handle barrier sectors
The market access API for barriers can return sectors as either a string
("All" or "N/A') or a list (of sectors). The string case wasn't handled
properly, resulting in e.g. an array like this: ["N", "/", "A"]. This
fix will convert "N/A" responses to Nones (SQL NULL), and the "All" case
into an array with one element: ["All"].