Split DB table cleanup operators into swap and temp table one
Cleaning up both temporary and swap tables at the end of the DAG
means we need to rerun insert data task even if pipeline fails on
swap or "check-temp-data" tasks (for example due to a misconfiguration
or a DB lock timing out the swap task).

Fetch and insert tasks are expensive and can take hours to finish, so
we want to avoid rerunning them if the succeed.

The new DAG structure has 2 cleanup tasks. Swap table cleanup runs
at the end and removes the 'swap' table containing previous dataset
version. It only runs if all other tasks succeeded.

Temp table cleanup task runs if any of the fetch, create table or
insert into temp table tasks have failed. Since in this case there's
no reason to believe the data in temp table is correct we drop the
entire table and the pipeline needs to be rerun either entirely or
starting from the "create table" task.

Failing either the check-temp-table-data or swap-dataset-table tasks
will leave the temp table so it can be either cleaned up manually or
the failed tasks can be rerun.

This split also makes DAG runs report accurate success/failure status.