Use a separate pool for sensor tasks
Airflow sensors are tasks that wait for an event to happen (like another
DAG run finishing for example). The sensors themselves don't take a lot
of resources, but since they are tasks, they take up a slot and a celery
worker while they wait.

This creates a possibility of a deadlock - if we have more sensors than
the number of concurrent tasks we can run it's possible for sensors to
take up all the workers, prevent the actual tasks to get scheduled and
wait until the timeout since their dependencies can't run until
something terminates.

In order to avoid this we create a separate Airflow pool specifically for
sensors. This limits how many sensors we run in parallel, leaving lots
of space for actual work.

Airflow with Celery executor has two main ways to control concurrency:
using celery queues and/or using Airflow pools.

The number of currently running tasks will be limited by both the number
of celery workers available (which is determined by worker concurrency
multiplied by the number of worker instances, so currently in production
it's 8 instances * 12 worker processes per instance) and the Airflow
slots, which control which tasks are scheduled to run (by default
there's one pool with 128 slots). There are also configuration options
that control for example how many tasks can run in parallel for each
DAG and limit active runs of the same DAG, but they're not very useful
to us at the moment.

The main reason we want to limit concurrent celery workers is that some
of our tasks can require a lot of memory (like fetching large responses
for example), so having too many of them per instance can lead to tasks
being killed by OOM.

So one approach would be to scale up the workers and limit concurrency
of CPU/DB/memory intensive tasks using pools instead. But this requires
a better understanding of what our limits look like, so for now limiting
sensors is an easier approach.

As long as the pipeline are structured with sensors being the first to
run before any work kicks off and DAGs don't have long chains of
dependencies (for example with A waiting for B which is waiting for C)
the worst case scenario with this set up is that sensors waiting for a
very long task take up the slots, meaning all other jobs with
dependencies will have to wait for the slow job to finish. But this is
not worse than our current approach of offsetting the DAGs by the max
dependency runtime.