Add 10 minute statement timeout to DB table operators
DB table operations are usually quick, unless they get locked by
a long running transactions, in which case they can get stuck for
hours without alerting anyone.

Airflow `execution_timeout` that we set on these tasks doesn't seem
to work for this specific case. It will interrupt a sleep or a long
running operation, but doesn't seem to be able to interrupt a locked
DB transaction, so we set an additional statement timeout on these.