Update UK trade in services pipelines
Changes:
* Don't join on period for the label, as the period has two labels for
reason and this was leading to duplicated data.
* Include the period types (quarter or year)

These pipelines will probably fail when this is deployed, as the new
pipeline will pull in less data. We'll need to manually delete the
existing tables so that the comparison doesn't happen. This should be OK
as the data is not currently published for users.