Increase page size for interactions dataset scrape
We scrape around 2.9M records of the interactions dataset daily, which
takes about 3 hours. We do this in chunks of 100, which feels pretty
inefficient. A PR on the data-hub-api
(https://github.com/uktrade/data-hub-api/pull/2749) allows clients to
specify their own page size, which will mean we can scrape in larger
chunks. This should reduce the total time-to-scrape as well as the time
to insert data into the DB (currently about 3.5 hours). Given a lot of
pipelines rely on this one, the faster it completes, the faster the
other pipelines can start and finish.