Timeout requests after 5 minutes
We keep waking up to the interactions dataset having hung for hours
trying to fetch data. By adding timeouts on the requests we're hoping to
enable them to get retried if they do get stuck for a little while.

This also adds retries to the fetch task for the interactions pipeline
as a whole, so that if the entire task dies it will automatically retry.