Setup a dedicated worker for 'heavy' tasks
ONS dataset 2 is a large dataset that needs to be pulled into memory all
at once when scraping and transforming data from the source XLSX file.
This is unlikely to run successfully on our standard 2GB workers, so
we'll set up a separate queue which will be processed by a new class of
worker: 'heavy'. We can scale this as we need, but will start off
probably with 4-5G RAM, which should be sufficient for this dataset.