Add dependencies to daily and monthly CSV pipelines
This replaces the time offset based scheduling of CSV pipelines with
task dependencies on all related dataset pipelines. The pipelines
will all start at the same time as the dataset ones, but will wait
form them to continue using sensor tasks.

This will make sure that any failed pipeline will also fail any
upstream CSV pipelines, so that they are clearly marked and can
be rerun together.

This adds a relatively large number of sensor tasks to the Airflow
pool, so we'll need to check that this doesn't slow down / block
tasks in production and potentially tweak the pool sizes based on
that.