Add Covid-19 Johns Hopkins University global time series data
The 3 CSVs published by CSSE are combined into one table using a
new operator that iterates over a mapping of source_urls and adds
the URL key as a "type" column to the CSV data, allowing us to
combine datasets that are published as separate files (for example
separate import and export data CSVs) into one table, which is
easier to use in the upcoming tools (since they don't generally
allow UNION queries).

The data is transformed from a wide format (with a column for each
date) into a normalized `(location, date, value)` list of measurements.

The new operator is using pandas to transform data, which means the
entire dataset is loaded into memory. This is not an issue for the
CSSE data, since it's relatively small (~60k rows), but might not
be appropriate for larger datasets without some changes.