Always keep S3 data for latest DAG run
Our S3 maintenance job cleans up old files used to store data during the
'fetch from api' phase of a pipeline. By default it drops data older
than 7 days. Some of our pipelines run monthly. Airflow pipelines record
their execution date related to the start of the period they're meant to
run for, which means that for monthly jobs the datetime tag on a
pipeline is always a month old - which means when our maintenance job
scans S3 files, it sees a date from a month ago and elects to clean it
up.

This patch forces our S3 maintenance job to always retain the latest
copy of the DAG run, regardless of how old it is.