Only drop temp tables if there is a table config
The cleanup task `drop_temp_tables` is there to remove temp tables when
a pipeline is complete. For the DSS pipelines, the first stage of the
pipeline is trying to retrieve the table config. If that fails for some
reason, when the cleanup task comes along it fails to read the table
config to know what to drop. So let's just check that the table config
exists in the first place.