URL quote schema/table names
Data Store Service has been accepting invalid DB identifiers for schemas
and table names (e.g. with spaces in) and sending these to Data Flow.
Data Flow then queries the Data Store Service again to read the table
config. When forming the URL, it puts the schema and table name in
without url encoding it, so when a space is present in the schema name
this isn't changed to %20 by Data Flow. DSS is protected by hawk
authentication, which is a signature formed from the URL and possibly
some other metadata. Data Flow therefore creates the hawk signature from
the non-encoded URL, but our request handler is smart enough to encode
the spaces to %20 before sending it off. When Data Store Service
receives this request and validates the signature, then, the URL doesn't
match the signature Data Flow constructed, and rejects the request with
a 401.

While the problem here originates from Data Store Service, and a
separate fix will go out to hopefully stop it sending us invalid db
identifiers in the first place, it would be good to fix this gap as
well.