feat: add metadata table
This adds alembic to data-flow so that it can set up and manage a
metadata table for data-flow pipelines/tables. In the beginning we will
only track the table schema, table name, source last modified date, and
pipeline table swap date. Where a pipeline does not explicitly define a
method to determine the source data modified date, we leave the
corresponding column in the metadata column empty for now. In the future
we are likely to want to fill this in. I'm choosing not to do this now
as it isn't clear on the correct timestamp to store here for real-time
systems, which most of the remaining pipelines will represent.

Note: this patch runs DB migrations as part of app startup, which isn't
best practice, though is what we generally do on our other apps at the
moment. I've made a ticket in the tech debt backlog to address this
anti-pattern.