feat: proof-of-concept tensorflow pipeline with Jenkins deployment
This is a whole separate Docker application since a buildpack
application with tensorflow doesn't seem to be possible due to a
CloudFoundry bug
https://github.com/cloudfoundry/cloud_controller_ng/issues/1802

We don't move all of AirFlow to Docker since we try to minimise its use
in PaaS.

The ci-pipeline only respects manifest.yml, but it is configured with a
pre-deploy command to copy manifest-tensorflow.yml to manifest.yml
https://github.com/uktrade/ci-pipeline-config/blob/master/datasci-data-flow-worker-tensorflow.yaml

The worker is, slightly misleadingly, the "web" process in the
application. Technically a "worker" Docker process is possible in
CloudFoundry via various entries in the manifest.yml, but the
ci-pipeline doesn't support specifying those entries.

The sample.env change for local development: I suspect previously, using
the local executor meant jobs weren't actually being sent via Redis and
run on the worker. It didn't make too much of a difference since the
environments were the same, but it's now vital to get jobs to run on the
Docker tensorflow worker.

For now, this is separate pipeline from Jenkins, which does mean it has
to be deployed separately. It could be integrated with the main
pipeline, but leaving it for now.