feat: send emails when a pipeline succeeds
We are starting to onboard some datasets that users want to start
working with as soon as possible after the source data has been updated.
In aid of this we need a mechanism to notify users when data has been
updated so that they don't need to keep checking themselves.

We will use GOV.UK Notify to send templated emails when a pipeline
succeeds which contain a link to the relevant Data Workspace dataset
page as the MVP.