fix: bypass scraper caching
The gss-utils scraper, by default, implements file-based caching for
requests. When we run our polling pipeline task in air-flow, we
initialise a new scraper each time we want to check the last modified
date but it resolves to the same file cache, meaning that we don't
actually retrieve the last data every time we poll. Providing a custom
session overrides the default caching logic meaning that every poll will
check for the latest date.