feat: use raw copy to load data
Rather than using pandas.DataFrame.to_sql to load data from the
dataframe to the DB, we instead rely on Postgres's native copy_from
functionality to provide a significant increase in data load times. In
order to do this we need to create the table ourselves, and this
actually brings us more in line with the standard pipeline where we now
declare our table columns in a (Single)TableConfig object up front, and
manipulate the resulting data into the table using source/target field
names.