fix: multiple queues on a single worker
We were trying to run two airflow workers on a single instance, one
listening to the default queue and another to the high-memory-usage
queue. This doesn't work - each airflow worker tries to listen on a
specific port, so the second worker dies with 'address already in use'
error.

To fix this, and maintain reasonable parity with our production setup,
we'll run a separate container with a worker that listens on the
high-memory-usage queue.