feat: longer waits and more retries for Activity Stream (actually this time)
Increasing number of attempts and delay between them since if a pipeline
fails, suspect the first thing that will be done in most cases would be
to retry, so we might as well have more retrying up front.

This was initially attempted in
3ec893d76be97319157bae0ca3e5ac248c9cad46, but that changed the wrong
_hawk_api_request function.

Leaving potentially merging the two functions until later.