feat: avoid storing all JSON in memory when loading from S3
In some cases, we load JSON manually into S3, and in some of those, the
JSON is larger than 1GB. Before this commit, this data is loaded into
memory at least twice, and most workers don't have enough memory for
this.

We could bump the memory, but in general, I think we should be making
small moves to reduce memory usage/support streaming, in order to
support larger datasets that can't be loaded into memory on a single
machine.

The use of the function insert_csv_data_into_db was removed in
2792f69f18c24e6741fea83f9bdbccab11ab28ee, and this was the only use of
json=False/jsonify=False, so opting to remove the function and these
attributes.