feat: use less memory in Companies House pipelines
The result of archive.open is a ZipExtFile, and its readlines method
reads all of its lines into a list.

However, ZipExtFile extends from BufferedIOBase
https://github.com/python/cpython/blob/3.7/Lib/zipfile.py#L785 which
extends from IOBase
https://github.com/python/cpython/blob/3.7/Lib/io.py#L78 and from
https://docs.python.org/3/library/io.html#io.IOBase

> IOBase (and its subclasses) supports the iterator protocol, meaning
> that an IOBase object can be iterated over yielding the lines in a
> stream.

so we can just avoid the readlines method, and use the ZipExtFile
directly.

It would be great if we could avoid loading the raw zip file in memory,
but AFAIK the directory for zip files is at the _end_ of the file, so I
don't think it's possible to stream-download a zip while unzipping its
contents.