fix: integrate tensorflow worker
We were using a separate app for the tensorflow worker previously
because GOV.UK PaaS didn't support customisable disk-space during
staging apps, and installing tensorflow caused staging to fail due to
out-of-disk errors. This has now been fixed, so we can use a standard
additional process for the tensorflow worker and drop the need for a
separate app (which is currently broken for deployment due to an
apparent incompatibility between kaniko and cloudfoundry:
https://github.com/cloudfoundry/grootfs/issues/71)

We have a baseline set of requirements that get installed for all
instances during the CloudFoundry staging process. In order to minimise
the amount of disk used by this stage, and required by all instances as
a minimum, we install tensorflow on only the tensorflow workers during
the instance startup step. If we installed tensorflow during staging,
all the apps end up needing about 1.5G disk each to hold the
installation. By installing tensorflow late, we can keep the majority of
instances on a lower disk size and reduce the size of the CloudFoundry
droplet. 6G is the maximum disk that GOV.UK PaaS allows to be allocated
to an instance (or staging), and when tensorflow was installed during
staging we were exceeding 5G, leading me to expect us to hit 6G
relatively soon, and therefore break severely again (unable to stage
apps, which means no deployments).