fix: set end dates for services/goods data
Comtrade API doesn't seem to have data for services before 2000 or for
goods before 1988. In order to get the pipeline working at all, let's
skip older years for now. We can revisit this if there is a pressing
need for the data or we discover that it should be there and we've
done something wrong.