fix: build payload for each request
The People Finder API is authenticated using JWT, which has an expiry
time that is validated by the server. At the moment we are setting the
expiry time as 60 seconds from the start of the pipeline, and re-using
the same payload/auth token for every subsequent request, meaning that
if the pipeline takes more than a minute we start to see failures.

We should generate a new token with a short(er) expiry for each request,
so that the pipeline can take an arbitrary amount of time to run and we
won't send expired tokens.