feat: support index creation after data insertion
For larger datasets, having the index established on the column before
we start inserting data leads to significant loss of efficiency/speed.
This patch provides support for defining indexes outside of the
`sa.Column` objects in the field mapping, meaning that columns can be
initially created without columns.

If any columns are defined on `TableConfig.indexes`, then we insert
an additional task into the DAG after the insert task to create the
indexes.

For pandas pipelines, we inline the post-insert-index-creation task
straight into the fetch-and-insert combined task.