fix: tweak task flow for pandas pipeline
The pandas pipeline currently merges the tasks for fetching data,
transforming it, ingesting it, and validating it into a single task. Now
that we are also building indexes after inserting the data, it feels
reasonable to split this task and data validation out into separate
steps. This pipeline is being used for some large datasets, and if the
index creation or data validation fail for transient reasons, we don't
want to have to re-fetch and insert all of the data again as it will
introduce significant delays.