fix: allow triggering DAGs from other DAGs
Our ONS pipelines come in two parts - polling pipelines that update
master dataset tables, and CSV generators that use the master datasets
to write out the latest data to CSV. The CSV pipelines use the
`dependencies` attribute to set up external task sensors on the polling
pipelines checking for a successful run of the `swap-dataset-tables`
task. This task only runs when the polling pipeline detects new data is
available, otherwise after a period of polling that task (and all
others) are skipped for the day. The `dependencies` framework ignores
skipped tasks.

The problem with this is that the external task sensor, on days where
new data isn't available, ends up timing out because it doesn't find a
successful run of the `swap-dataset-tables` task. This throws errors,
which get reported in sentry. We don't want these errors, because this
is OK behaviour in this instance.

What we need is to have the CSV pipelines run when the polling pipeline
has pulled in new data, and not run/error if it doesn't.

I considered a few solutions:

1) Turn off scheduled runs of the CSV pipeline and trigger it manually
   when the polling pipeline has pulled in new data. This is sort of the
   opposite of the `dependencies` process. It also only works when there
   is a single parent pipeline, which in our current ONS example is
   true. If there were multiple parents, triggering the CSV pipeline
   manually would not be feasible.
2) Change the external task sensor to accept skipped
   `swap-dataset-tables` tasks, perhaps only for specific pipelines.
   This would then result in the CSV being generated again, redundantly,
   every day. This would also mean that the sensors are running
   throughout the day taking up a slot for limited reason.
3) Sub-class the ExternalTaskSensor in some way, for specific pipelines,
   to suppress the error that gets thrown. This would again mean the
   sensor still runs a lot, taking up a slot, but then when it dies we
   don't get noisy errors where we don't want them.

This patch goes for option 1 as a relatively simple implementation that
reduces the number of sensors running for little reason.